refactor props AccordeonBox.menu
