+++ services/app/assets/js/widgets/components/ExecutionOutput/ExecutionOutput.jsx
@@ -1,20 +1,6 @@
 import React from 'react';
-import i18n from '../../../i18n';
 import AccordeonBox from './AccordeonBox';
 import statusColor from '../../config/statusColor';
-
-const getMessage = status => {
-  switch (status) {
-    case 'error':
-      return i18n.t('You have some syntax errors');
-    case 'failure':
-      return i18n.t('Test failed');
-    case 'ok':
-      return i18n.t('Yay! All tests passed!!111');
-    default:
-      return i18n.t('Press Check solution or press Give up');
-  }
-};
 
 const parseResult = result => {
   if (result) return JSON.parse(result);
@@ -29,19 +15,13 @@ const ExecutionOutput = ({
   const resultData = parseResult(result);
   const allAsserts = asserts.map(JSON.parse);
   const [firstAssert, ...restAsserts] = allAsserts;
-  const percent = (100 * successCount) / assertsCount;
-  const isSyntaxError = resultData.status === 'error';
   return (
     <AccordeonBox>
       <AccordeonBox.Menu
-        count={i18n.t(
-          'You passed %{successCount} from %{assertsCount} asserts. (%{percent}%)',
-          { successCount, assertsCount, percent },
-        )}
-        statusColor={statusColor[resultData.status]}
-        message={getMessage(resultData.status)}
+        assertsCount={assertsCount}
+        successCount={successCount}
         firstAssert={firstAssert}
-        isSyntaxError={isSyntaxError}
+        resultData={resultData}
       >
         {resultData.status === 'error' ? (
           <AccordeonBox.Item
